requirements.txt
langchain>=0.1.0
langchain-community
arize-phoenix[evals]
nest-asyncio
pyarrow
python-docx
pydantic
phoenix
requests
pyprojroot
openinference-instrumentation-langchain
langchain-openai
opentelemetry-sdk
opentelemetry-api
openpyxl
uv
mcp
mcp[cli]